+++ test/conftest.py
@@ -1,0 +1,5 @@
+import os
+
+
+def pytest_configure():
+    os.environ["AWS_EC2_METADATA_DISABLED"] = "true"