+++ test/test_prod_handler.py
@@ -1,18 +1,33 @@
+# The MIT License (MIT)
+# Copyright (c) 2025 by the xcube development team and contributors
+#
+# Permission is hereby granted, free of charge, to any person obtaining a
+# copy of this software and associated documentation files (the "Software"),
+# to deal in the Software without restriction, including without limitation
+# the rights to use, copy, modify, merge, publish, distribute, sublicense,
+# and/or sell copies of the Software, and to permit persons to whom the
+# Software is furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NON INFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+
 import unittest
 from unittest.mock import Mock, patch
 
-from xcube.core.store.preload import NullPreloadHandle
-from xcube.util.jsonschema import JsonObjectSchema
-
-from xcube_clms.constants import (
-    CLMS_DATA_ID_KEY,
-    DATASET_DOWNLOAD_INFORMATION,
-    FULL_SOURCE,
-    ITEMS_KEY,
-)
+from xcube_clms.constants import (CLMS_DATA_ID_KEY,
+                                  DATASET_DOWNLOAD_INFORMATION, FULL_SOURCE,
+                                  ITEMS_KEY)
 from xcube_clms.product_handler import ProductHandler
+from xcube_clms.product_handlers.cdse import CdseProductHandler
 from xcube_clms.product_handlers.eea import EeaProductHandler
-from xcube_clms.product_handlers.legacy import LegacyProductHandler
 
 
 class TestProductHandler(unittest.TestCase):
@@ -27,7 +42,7 @@ class TestProductHandler(unittest.TestCase):
             },
             {
                 CLMS_DATA_ID_KEY: "daily-surface-soil-moisture-v1.0",
-                DATASET_DOWNLOAD_INFORMATION: {ITEMS_KEY: [{FULL_SOURCE: "legacy"}]},
+                DATASET_DOWNLOAD_INFORMATION: {ITEMS_KEY: [{FULL_SOURCE: "cdse"}]},
             },
         ]
 
@@ -48,29 +63,14 @@ class TestProductHandler(unittest.TestCase):
         )
         self.assertIsInstance(handler, EeaProductHandler)
 
-    def test_guess_legacy(self):
+    def test_guess_cdse(self):
         handler = ProductHandler.guess(
             data_id="daily-surface-soil-moisture-v1.0",
             datasets_info=self.datasets_info,
             cache_store=self.cache_store,
             credentials=self.mock_credentials,
         )
-        self.assertIsInstance(handler, LegacyProductHandler)
-
-    def test_legacy_preload(self):
-        handler = ProductHandler.guess(
-            data_id="daily-surface-soil-moisture-v1.0",
-            datasets_info=self.datasets_info,
-            cache_store=self.cache_store,
-            credentials=self.mock_credentials,
-        )
-        preload_handle = handler.preload_data(("daily-surface-soil-moisture-v1.0",))
-        self.assertIsInstance(preload_handle, NullPreloadHandle)
-
-        schema = handler.get_preload_data_params_schema()
-        self.assertEqual(
-            vars(JsonObjectSchema(additional_properties=False)), vars(schema)
-        )
+        self.assertIsInstance(handler, CdseProductHandler)
 
     def test_guess_missing_args_raises(self):
         with self.assertRaises(ValueError) as context: